test(otel): stop guessing a dead port
`dead_endpoint()` was the last bind-then-drop after #173 and #222: bind
:0, read the number, drop the listener, point the OTLP exporter at it.
The exposure is the inverse of the handoff race — this one wants the port
dead. A stranger that takes it and answers 2xx keeps the export healthy,
so `wait_failing(true)` never trips.

The issue calls that a quiet false pass; it is not. Reproduced by binding
the released port with a squatter answering `200 OK`, the test dies at
the barrier — "the audit sink did not reach failing()=true within 8s" —
before any outage assertion runs. Loud, like #173 and #222, but the panic
names the audit sink and not the port collision, so the diagnosis it
offers is a wrong one. Only the leading "open keeps serving" assertion
passes vacuously, and a healthy collector satisfies that too.

Fix is `refused_base_url()`, the in-repo precedent, which already
documents bind-then-drop as the pattern it avoids: `127.0.0.1:1` is
privileged, so no non-root test process can occupy it, and `resolve()`
appends `/v1/logs` to it as it does to any endpoint. It moves to
`common/refused.rs` because `otel_wiremock.rs` does not say `mod common;`
and the sibling `REFUSED_CONNECT_BUDGET` would be `dead_code` there under
`-D warnings` (#167, #214); the two existing callers take it by `#[path]`
on the `raw_post.rs` / `pinned_cli.rs` pattern. No assertion changed.

Fixed form: the same squatter cannot be constructed against it (EACCES on
127.0.0.1:1), and 40 runs of the suite under a churner holding 2000
loopback ports answering 200 gave 0 failures.

Closes #229
