+++ crates/bugwarden/tests/preflight_wiremock.rs
@@ -31,6 +31,8 @@ use wiremock::{Mock, MockServer, ResponseTemplate};
 
 #[path = "common/pinned_cli.rs"]
 mod pinned_cli;
+#[path = "common/refused.rs"]
+mod refused;
 
 use pinned_cli::pinned;
 
@@ -277,7 +279,7 @@ async fn preflight_transport_error_does_not_leak_the_api_key_i12() {
     // the whoami lookup fails at the transport level, where the
     // unsanitized error would carry the request URL with api_key=... in
     // it. Nothing in the preflight error text may contain the key.
-    let base = common::refused_base_url();
+    let base = refused::refused_base_url();
 
     let cfg: Arc<Cli> = Arc::new(pinned(&[
         "bugwarden",
